refactor: 인터페이스에 객체가 드러나도록 변경

+++ src/main/kotlin/lotto/TicketAmount.kt
@@ -2,5 +2,5 @@ package lotto
 
 class TicketAmount(count: Int = 1) {
     val money: Money = Money(1_000 * count)
-    val amount: Int = money.amount
+    val amount: Long = money.amount
 }
